# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CoopLumen API
  description: REST API for CoopLumen communities and Stellar assets.
  version: 0.1.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v1/balances/{publicKey}:
    parameters:
      - name: publicKey
        in: path
        required: true
        description: Stellar public key of the account to inspect.
        schema:
          $ref: '#/components/schemas/StellarKey'
    get:
      tags: [Balances]
      summary: Get account balances
      description: >
        Returns all on-chain balances for a Stellar account. Successful responses
        may be served from a Redis cache for up to 5 seconds; the response body
        itself is unchanged.
      responses:
        '200':
          description: Account balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502':
          description: Horizon is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/balances/{publicKey}/loans:
    parameters:
      - name: publicKey
        in: path
        required: true
        description: Stellar public key to look up as either borrower or lender.
        schema:
          $ref: '#/components/schemas/StellarKey'
    get:
      tags: [Balances, Loans]
      summary: List loans involving an address
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of loans where the address is borrower or lender
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Loan' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /api/v1/balances/community/{communityId}/loans:
    parameters:
      - $ref: '#/components/parameters/CommunityId'
    get:
      tags: [Balances, Loans]
      summary: List loans in a community
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of loans belonging to the community
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Loan' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
components:
  parameters:
    CommunityId:
      name: id
      in: path
      required: true
      description: Community UUID.
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      description: Page size, capped at 100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    StellarKey:
      type: string
      description: A 56-character Stellar StrKey public key.
      pattern: '^G[A-Z2-7]{55}$'
    BalanceLine:
      type: object
      properties:
        asset_type: { type: string }
        asset_code: { type: string, nullable: true }
        asset_issuer: { type: string, nullable: true }
        balance: { type: string }
    BalanceResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BalanceLine'
...
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }
        offset: { type: integer }
    ValidationIssue:
      type: object
      properties:
        path:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        message: { type: string }
    ValidationErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
        meta:
          type: object
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationIssue'
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
    Loan:
      type: object
      properties:
        id: { type: string, format: uuid }
        community_id: { type: string, format: uuid }
        borrower_address: { $ref: '#/components/schemas/StellarKey' }
        lender_address: { $ref: '#/components/schemas/StellarKey' }
        amount: { type: string }
        amount_repaid: { type: string }
        asset_code: { type: string }
        asset_issuer: { $ref: '#/components/schemas/StellarKey' }
        purpose: { type: string, nullable: true }
        status:
          type: string
          enum: [pending, active, repaid, defaulted, cancelled]
        due_at: { type: string, format: date-time, nullable: true }
        disbursed_at: { type: string, format: date-time, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...